# docs/swagger.yaml
basePath: /chaos/api/v1
definitions:
...
  recover.Alert:
    properties:
      labels:
        $ref: '#/definitions/recover.Labels'
      status:
        type: string
    type: object
  recover.Labels:
    properties:
      recoverAll:
        type: boolean
      recoverJob:
        type: string
      recoverTarget:
        type: string
    type: object
  recover.RequestPayload:
    properties:
      alerts:
        items:
          $ref: '#/definitions/recover.Alert'
        type: array
    type: object
...
  response.RecoverMessage:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  response.RecoverResponsePayload:
    properties:
      recoverMessages:
        items:
          $ref: '#/definitions/response.RecoverMessage'
        type: array
      status:
        type: integer
    type: object
...
host: localhost:8090
info:
  contact: {}
  description: This is the chaos master API.
  title: Chaos Master API
  version: "1.0"
paths:
...
  /recover/alertmanager:
    post:
      consumes:
      - application/json
      description: Alertmanager webhook to recover from failures
      parameters:
      - description: Create request payload that contains the recovery details
        in: body
        name: RequestPayload
        required: true
        schema:
          $ref: '#/definitions/recover.RequestPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RecoverResponsePayload'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: recover from failures
      tags:
      - Recover
...
swagger: "2.0"